Do not crash after connection is closed.
In cases where there is a connection timeout set in the redis.conf,
we do not crash.

Basically it has been handled by using the retry flag on the node-redis.
we are also not hooking to the end event on the connection to track the
server up/down status. This is as previously handled by the ping.

There may also be problems if the ping timeout is set to a higher number
than the connection timeout.

We recommend using a ping timeout much lesser than redis connection timeout.